refactor: 상품 도메인 계층 패키지 변경
