+++ src/main/java/shop/tryit/domain/cart/entity/CartItem.java
@@ -13,7 +13,7 @@ import javax.persistence.ManyToOne;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.tryit.domain.item.Item;
+import shop.tryit.domain.item.entity.Item;
 
 @Entity
 @Getter
