+++ src/main/java/shop/tryit/domain/cart/repository/CartItemRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 import shop.tryit.domain.cart.entity.Cart;
 import shop.tryit.domain.cart.entity.CartItem;
-import shop.tryit.domain.item.Item;
+import shop.tryit.domain.item.entity.Item;
 
 public interface CartItemRepository {
