+++ src/main/java/shop/tryit/domain/item/dto/ItemDto.java
@@ -9,7 +9,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 import shop.tryit.domain.item.Category;
-import shop.tryit.domain.item.Image;
+import shop.tryit.domain.item.entity.Image;
 
 @Data
 @NoArgsConstructor(access = PROTECTED)
