+++ src/main/java/shop/tryit/domain/order/dto/OrderDetailSearchDto.java
@@ -2,7 +2,7 @@ package shop.tryit.domain.order.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
-import shop.tryit.domain.item.ItemSearchDto;
+import shop.tryit.domain.item.dto.ItemSearchDto;
 
 @Getter
 public class OrderDetailSearchDto {
