+++ src/main/java/shop/tryit/domain/order/service/OrderFacade.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
-import shop.tryit.domain.item.Item;
+import shop.tryit.domain.item.entity.Item;
 import shop.tryit.domain.order.Order;
 import shop.tryit.domain.order.OrderDetail;
 import shop.tryit.domain.order.dto.OrderDto;
