+++ src/main/java/shop/tryit/domain/payment/PaymentFacade.java
@@ -3,10 +3,10 @@ package shop.tryit.domain.payment;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
-import shop.tryit.domain.item.Image;
-import shop.tryit.domain.item.ImageService;
-import shop.tryit.domain.item.Item;
-import shop.tryit.domain.item.ItemService;
+import shop.tryit.domain.item.entity.Image;
+import shop.tryit.domain.item.entity.Item;
+import shop.tryit.domain.item.service.ImageService;
+import shop.tryit.domain.item.service.ItemService;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberService;
 import shop.tryit.domain.payment.dto.PaymentDto;
