+++ src/main/java/shop/tryit/repository/cart/CartItemJpaRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.tryit.domain.cart.entity.Cart;
 import shop.tryit.domain.cart.entity.CartItem;
-import shop.tryit.domain.item.Item;
+import shop.tryit.domain.item.entity.Item;
 
 public interface CartItemJpaRepository extends JpaRepository<CartItem, Long> {
