+++ src/main/java/shop/tryit/repository/item/ItemCustom.java
@@ -3,7 +3,7 @@ package shop.tryit.repository.item;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import shop.tryit.domain.item.ItemSearchCondition;
-import shop.tryit.domain.item.ItemSearchDto;
+import shop.tryit.domain.item.dto.ItemSearchDto;
 
 public interface ItemCustom {
