+++ src/main/java/shop/tryit/repository/item/ItemJpaRepository.java
@@ -1,7 +1,7 @@
 package shop.tryit.repository.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.item.Item;
+import shop.tryit.domain.item.entity.Item;
 
 public interface ItemJpaRepository extends JpaRepository<Item, Long>, ItemCustom {
