+++ src/main/java/shop/tryit/repository/item/ItemRepositoryImpl.java
@@ -7,10 +7,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
-import shop.tryit.domain.item.Item;
-import shop.tryit.domain.item.ItemRepository;
 import shop.tryit.domain.item.ItemSearchCondition;
-import shop.tryit.domain.item.ItemSearchDto;
+import shop.tryit.domain.item.dto.ItemSearchDto;
+import shop.tryit.domain.item.entity.Item;
+import shop.tryit.domain.item.repository.ItemRepository;
 
 @Slf4j
 @Repository
