+++ src/main/java/shop/tryit/web/item/ImageController.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseBody;
-import shop.tryit.domain.item.ImageStore;
+import shop.tryit.domain.item.service.ImageStore;
 
 @Slf4j
 @Controller
