+++ src/test/java/shop/tryit/domain/item/ItemServiceTests.java
@@ -14,6 +14,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
+import shop.tryit.domain.item.dto.ItemSearchDto;
+import shop.tryit.domain.item.entity.Item;
+import shop.tryit.domain.item.repository.ItemRepository;
+import shop.tryit.domain.item.service.ItemService;
 
 @Transactional
 @SpringBootTest
