+++ src/test/java/shop/tryit/domain/order/OrderServiceTests.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.domain.item.Item;
-import shop.tryit.domain.item.ItemRepository;
+import shop.tryit.domain.item.entity.Item;
+import shop.tryit.domain.item.repository.ItemRepository;
 import shop.tryit.domain.member.Member;
 import shop.tryit.domain.member.MemberRepository;
 import shop.tryit.domain.order.service.OrderService;
